feat: move logic code into lib file
Move all the logic into a separate lib file.
Now `main.rs` is only responsible for running the library code and holds no logic at all.
